Appointment features are done

+++ projeto-es/Presentation_Layer/UserMenuAppointmentsForm.cs
@@ -38,15 +38,17 @@ namespace projeto_es.Presentation_Layer
         { 
             
             var Appointments = appointmentService.GetsAppointmentsOfAClient(LoggedSession.Account.Id);
-
+            //appointmentsListBox.Items.Add("Name of Staff | scheduled time");
             foreach (var Appointment in Appointments)
             {
+                
                 appointmentsListBox.Items.Add(Appointment.description);
+                scheduledTimeListBox.Items.Add(Appointment.scheduled_time);
                 AppointmentIDlistBox.Items.Add(Appointment.id);
             }
         }
 
-        private void checkAvailability()
+        private bool checkAvailabilityCancel()
         {
 
             DateTime localDate = DateTime.Now;
@@ -67,29 +69,94 @@ namespace projeto_es.Presentation_Layer
 
                     if (selectedDate < localDate)
                     {
-                        Console.WriteLine("The selected date is not possible.");
+                        errorLabel.Text = "The selected date is an old appointment";
+                        return false;
 
                     }
                     else
                     {
-                        Console.WriteLine("The selected date good.");
+                        return true;
                     }
 
                 }
 
             }
+            return false;
 
         }
+
+        private bool checkAvailabilityAttend()
+        {
+
+            DateTime localDate = DateTime.Now;
+
+            var Appointments = appointmentService.GetsAppointmentsOfAClient(LoggedSession.Account.Id);
+
+
+            foreach (var appointment in Appointments)
+            {
+                if (appointment.id == (uint)AppointmentIDlistBox.SelectedItem)
+                {
+
+                    DateTime selectedAppointmentDate = Convert.ToDateTime(appointment.scheduled_time);
+                    DateTime toleranceAppointmentEarly = selectedAppointmentDate.AddMinutes(-10);
+                    DateTime toleranceAppointmentLate = selectedAppointmentDate.AddMinutes(10);
+
+                    //Console.WriteLine("localdate: " + localDate);
+                    //Console.WriteLine("selecteddate: " + selectedAppointmentDate);
+
+                    //Console.WriteLine("selectedDate.Date: " + selectedAppointmentDate.Date);
+                    //Console.WriteLine("localDate.Date: " + localDate.Date);
+
+                    //Console.WriteLine("ToleranceEarly: " + toleranceAppointmentEarly);
+                    //Console.WriteLine("ToleranceLate: " + toleranceAppointmentLate);
+
+
+                    if (selectedAppointmentDate.Date == localDate.Date)
+                    {
+                        if(localDate.TimeOfDay < toleranceAppointmentEarly.TimeOfDay)
+                        {
+                            errorLabel.Text = "It's too early to be attended";
+                            break;
+                        }
+
+                        if(localDate.TimeOfDay > toleranceAppointmentLate.TimeOfDay)
+                        {
+                            errorLabel.Text = "You lost the oportunity to be attended";
+                            break;
+                        }
+
+                        return true;
+
+                    }
+                    else
+                    {
+                        errorLabel.Text = "The selected date is not available";
+                        return false;
+                    }
+
+                }
+
+            }
+            return false;
+
+        }
+
         private void atendAppointmentButton_Click(object sender, EventArgs e)
         {
-            checkAvailability();
-
+            if (checkAvailabilityAttend())
+            {
+                Console.WriteLine("Ayyyy lmaaao");
+            }
         }
 
         private void cancelAppointmentButton_Click(object sender, EventArgs e)
         {
-            appointmentService.DeleteAppointment((uint)AppointmentIDlistBox.SelectedItem);
-            Console.WriteLine("Deleted appointment: " + (uint)AppointmentIDlistBox.SelectedItem);
+            if (checkAvailabilityCancel())
+            {
+                appointmentService.DeleteAppointment((uint)AppointmentIDlistBox.SelectedItem);
+                Console.WriteLine("Deleted appointment: " + (uint)AppointmentIDlistBox.SelectedItem);
+            }
 
         }
